closeOpenChats
